requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
requests==2.31.0
python-socketio==5.10.0

# Crypto & Financial Data
ccxt>=4.2.0
yfinance>=0.2.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
ta>=0.11.0
# pandas-ta - Install from GitHub: pip install git+https://github.com/twopirllc/pandas-ta

# Deep Learning & AI
torch>=2.0.0,<3.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# FinBERT for sentiment analysis
sentencepiece>=0.1.99

# Backtesting
backtrader>=1.9.78
# vectorbt>=0.26.0  # Optional - large dependency

# News & Sentiment
# newsapi-python>=0.2.7  # Optional
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Technical Analysis
# mplfinance>=0.12.0  # Optional - for plotting

# Database (Optional)
# sqlalchemy>=2.0.0
# redis>=5.0.0

# Utils
python-dotenv>=1.0.0
# schedule>=1.2.0  # Optional
python-dateutil>=2.8.0

# Visualization
plotly>=5.18.0
